Evaluate all formulas before saving an xssf document
Since we are using SXSSF (streamed xssf), it might be
that we can only refer to cells which are currently
in the buffer and that this implemenation does not
work for very large documents.
