+++ src/test/java/com/docutools/jocument/impl/excel/implementations/ExcelGeneratorTest.java
@@ -7,6 +7,7 @@ import com.docutools.jocument.TestUtils;
 import com.docutools.jocument.impl.ReflectionResolver;
 import com.docutools.jocument.sample.model.SampleModelData;
 import com.docutools.poipath.PoiPath;
+import com.docutools.poipath.xssf.XSSFWorkbookWrapper;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
@@ -156,7 +157,9 @@ class ExcelGeneratorTest {
         assertThat(firstSheet.row(3).cell(0).doubleValue(), closeTo(4.0, 0.1));
         assertThat(firstSheet.row(4).cell(0).doubleValue(), closeTo(5.0, 0.1));
         assertThat(firstSheet.row(7).cell(0).text(), equalTo("SUM(A1:A5)"));
+        assertThat(firstSheet.row(7).cell(0).intValue(), equalTo(15));
         assertThat(firstSheet.row(7).cell(1).text(), equalTo("COUNT(A1:A5)"));
+        assertThat(firstSheet.row(7).cell(1).intValue(), equalTo(5));
     }
 
     @Test
@@ -182,5 +185,24 @@ class ExcelGeneratorTest {
         assertThat(firstSheet.row(41).cell(1).stringValue(), equalTo("Exarcheia"));
         assertThat(firstSheet.row(42).cell(1).stringValue(), equalTo("Nova Metalkova"));
         assertThat(firstSheet.row(52).cell(0).stringValue(), startsWith("Das Denken"));
+    }
+
+    @Test
+    void shouldResolveHyperlink() throws InterruptedException, IOException {
+        // Arrange
+        Template template = Template.fromClassPath("/templates/excel/HyperlinkDocument.xlsx")
+            .orElseThrow();
+        PlaceholderResolver resolver = new ReflectionResolver(SampleModelData.PICARD);
+
+        // Act
+        Document document = template.startGeneration(resolver);
+        document.blockUntilCompletion(60000L); // 1 minute
+
+        // Assert
+        assertThat(document.completed(), is(true));
+        var xssfWorkbook = TestUtils.getXSSFWorkbookFromDocument(document);
+        var documentWrapper = new XSSFWorkbookWrapper(xssfWorkbook);
+        assertThat(documentWrapper.sheet(0).row(0).cell(0).text(), equalTo("orf.at"));
+        assertThat(documentWrapper.sheet(0).row(0).cell(0).cell().getHyperlink().getAddress(), equalTo("https://orf.at/"));
     }
 }
